Factor out sensitivity related functions
Also added a note on the use of gdk_pixmap_new in a kaera function which outputs
a warning about incompatible pointer types. It probably still works because all
the calls utilizing that pointer are EXPECTING that other type, which is also
why it breaks when I change it to some more modern cairo stuff. I need to
replace all those calls first.