Increase amount of displayed kanji
I know that with the current method of rendering, I can't actually display that
many kanji on the screen at once - however, just knowing how I CAN change the
amount of kanji which is actually suggested will be helpful for when I start
tweaking the UI to fit my needs better. For now, I set the value to 100, which
should be plenty.

Next up: Tweaking the UI in order to display all of these newfound kanji
suggestions!