Insert copyright notice into most source files
I modified the comment in the main file to make my goals more achievable, and
then pasted the copyright notice (along with a short notice of modification) in
all the source and header files in the main project.

This means that in theory, I now feel like I can properly fork the repo without
feeling like I'm not paying proper attribution to the involved parties. I still
have more things to commit, though, in order to get things up to speed, so I'll
be staving that part off for a little.

My ultimate goal with forking this repo to github is to get remote access to my
changes so I can clone the repo onto my RPi and see if I can get a "pocket
dictionary" up and working. However, I welcome anyone interested to take a look
at the project and see if you can use it for your own purposes. That is to say,
if anyone even ever reads this commit message...