Add jukugo search box, more results, scrolling, ++
Added a jukugo search box which will ultimately be used for querying jisho.org
using the RESTful jisho.org API. Also made it so that more results can be shown
when searching (this involved modifying both jstroke/jstroke.h and the
preprocessor definitions for number of results, as well as adding a workaround
for a resulting bug which arose due to the input apparently creating some input
to the engine which could not be converted to UTF8 or something like that... the
workaround, btw, is to just display a filler character (maru) instead.). I also
made it possible to scroll (and to add characters to the search box by either
double clicking them or pressing Ctrl+A when it's selected). Furthermore, I
fixed a crash arising from "undoing" when no strokes had been registered.