Add skeleton for displaying results function
Made a header declaring the function and a source file containing it. Also
updated the makefile to include the new object file in the build, so everything
should be fine.

For the time being it prints the results to the terminal just like before, but I
will change that in a later commit. I've added a note to explain what I intend
to do (and some things I don't).

TODO: include copyright notice ++ for consistency.