Add additional setup for the difficult part
I included YAJL in the json_to_gtk source file and, fixing a resulting build
error, adding a -I. to the include flags in the makefile. That's actually an
issue on YAJL's part, but whatever.

I also created a global vbox variable in which to put the results, and made it
accessible from json_to_gtk. Time will show if this naïve approach will work.