Further develop results display
Mainly, the changes are: I added a scrollbox and I'm setting its height (the
size of which is defined in preprocessor_definitions.h), which again contains
the vbox which I'm adding entries to. The vbox is cleared when display_results
is called and the scrollbox with content is hidden upon clearing the search
entry field. I'm pretty pleased with how things are right now. This is almost
release-worthy! (Or not.)