Fix an error which failed RPi build
An error happened at jisho_searc.c:75:46 because "'struct hostent'
has no member named 'h_addr'". I saw somewhere that h_addr is just
defined to be h_addr_list[0] "for backward compatibility", so I changed
it to just say that instead.

Also update gitignore to not include yajl in the project. It's better if
it's built on the target computer.