Fix bugs causing crash
Case in point: Some times, an entry does not have an "is_common" element, nor
does it always have a reading. The solution was to change the calls to yajl_get
into yajl_get_soft and handle the unexpected result more gently. Now, if there
is no reading, the program just doesn't display the reading. (I also took the
liberty of not displaying the reading as a reading if there is no kanji, since
then we just display the reading by default anyways.) Also, if both the word and
the reading are absent, the program doesn't crash, it merely prints an error and
skips. (This might lead to unintended results, but not to hard failures, which
is kinda more important at this stage. I can't control the works of the API.)
Furthermore, if is_common is absent, it's interpreted as "not common".