Start working on #4 - virtual keyboard
I added a "Toggle keyboard" menu item. It doesn't _do_ anything yet, but it's
there nonetheless, and the callback function is technically existing. I'll have
to decide whether to make a floating thing or to make it occupy the same space
as the results or something. A floatie would probably be better, but an internal
thing might be _easier_...