Create keyboard structure (#4)
Created a Keyboard structure which contains whether keyboard is visible and the
widget containing the keys - this may be changed in the future. Also created a
new file keyboard.c and its corresponding header, in which the function
keyboard_set_visible is declared and defined. The callback function merely calls
this function.