Fix #4 - virtual keyboard
Virtual keyboard with hiragana replacement is complete. Although not all
possible combinations are encoded - notably, the te-xi and te-xu combinations, as
well as the fu-xi, fu-xe, fu-xa, fu-xo combinations, have not been added.

I'm way too tired to keep going tonight, so this will have to do. I'm pretty
pleased with how it turned out in the end!