Fix #12 - keyboard shows/hides as per focus
I had to abandon the idea that the keyboard shows if the entry receives focus,
as that seems to be the default mode for the toplevel window. However, I managed
to make it display whenever a mouse button is clicked on the jukugo_entry
widget, and it hides when the keyboard window loses focus. So this should be all
right.