Begin working on #10 - incomplete hiragana
I've fixed all the problems mentioned in the initial description, aside from the
use of xke and xka, but I've yet to come up with a convincing solution to the
"nn" vs "n" problem.

NOTE: The initial description mistakingly wrote that てぅ should be used for
"tu", whereas the standard way to do it seems to be via とぅ, which is what I've
done now. I've also made it so that "di" and "du" are no longer synonymous with
"dji" and "dzu" - for the latter, I simply removed the previous "du" map, since
"dzu" was already a map to づ, but for the former I had to change the mapping
from "di" to "dji". However, that one is used less than じ by far, so this
non-standard writing is unlikely to be an issue. For the most part, the hiragana
translation should be working now.