Fix #10 - incomplete hiragana
I decided to let "nn" translate to "んn", and to let the single hiragana ん be
mapped from "n " (n and a space). In order to allow for input of single ん even
on a computer without a physical keyboard (read: RPi with just a touch screen),
I added a spacebar covering the entire row below the last row containing
letters, and increased the size of the popup window so it would show. This may
still not be ideal, but it's good enough!